pvr.iptvsimple fix

+++ script.maintenance.management/main.py
@@ -105,14 +105,14 @@ def cleanuserdata():
 	sizeMB = 0
 	dir,files = xbmcvfs.listdir(userdatafolder)
 	for directories in dir:
-		if not 'service.openelec.settings' in directories:
+		if not 'service.openelec.settings' in directories and not 'pvr.iptvsimple' in directories:
 			if not xbmcvfs.exists(os.path.join(addonsfolder,directories+'/')):
 				textmsg = textmsg+directories+'\n'
 				sizeMB = sizeMB + returnsize(os.path.join(userdatafolder,directories))
 	ok = mensagemyesno('Pastas a serem eliminadas do userdata:',textmsg+str(round(sizeMB,2))+' MB\nDeseja Continuar?')
 	if ok and textmsg:
 		for directories in dir:
-			if not 'service.openelec.settings' in directories:
+			if not 'service.openelec.settings' in directories and not 'pvr.iptvsimple' in directories:
 				if not xbmcvfs.exists(os.path.join(addonsfolder,directories+'/')): shutil.rmtree(os.path.join(userdatafolder,directories+'/'))
 		ok = mensagemok('Concluido','Operação Terminada')
